integration/requirements.txt
cattle==0.5.1
directio==1.2
protobuf==3.8.0
grpcio==1.21.1

flake8
pytest==5.3.2
pytest-parallel==0.0.10

libqcow-python==20181227
